shift all jdk11 tests to their own package
