+++ com.ibm.wala.core/src/test/java/com/ibm/wala/core/tests/jdk11/nestmates/NestmatesTest.java
@@ -1,4 +1,14 @@
-package com.ibm.wala.core.tests.stringConcat;
+/*
+ * Copyright (c) 2006 IBM Corporation.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ */
+package com.ibm.wala.core.tests.jdk11.nestmates;
 
 import com.ibm.wala.core.tests.callGraph.CallGraphTestUtil;
 import com.ibm.wala.core.tests.util.WalaTestCase;
@@ -14,7 +24,7 @@ import java.io.IOException;
 import org.junit.Assert;
 import org.junit.Test;
 
-public class StringConcatTest extends WalaTestCase {
+public class NestmatesTest extends WalaTestCase {
   @Test
   public void testPrivateInterfaceMethods()
       throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
@@ -24,7 +34,7 @@ public class StringConcatTest extends WalaTestCase {
     ClassHierarchy cha = ClassHierarchyFactory.make(scope);
     Iterable<Entrypoint> entrypoints =
         com.ibm.wala.ipa.callgraph.impl.Util.makeMainEntrypoints(
-            scope, cha, "LstringConcat/StringConcat");
+            scope, cha, "Lnestmates/TestNestmates");
 
     AnalysisOptions options = CallGraphTestUtil.makeAnalysisOptions(scope, entrypoints);
 
@@ -33,21 +43,26 @@ public class StringConcatTest extends WalaTestCase {
 
     // Find node corresponding to main
     TypeReference tm =
-        TypeReference.findOrCreate(ClassLoaderReference.Application, "LstringConcat/StringConcat");
+        TypeReference.findOrCreate(ClassLoaderReference.Application, "Lnestmates/TestNestmates");
     MethodReference mm = MethodReference.findOrCreate(tm, "main", "([Ljava/lang/String;)V");
     Assert.assertTrue("expect main node", cg.getNodes(mm).iterator().hasNext());
     CGNode mnode = cg.getNodes(mm).iterator().next();
 
-    // should be from main to testConcat()
+    // should be from main to Triple()
     TypeReference t1s =
-        TypeReference.findOrCreate(ClassLoaderReference.Application, "LstringConcat/StringConcat");
-    MethodReference t1m = MethodReference.findOrCreate(t1s, "testConcat", "()Ljava/lang/String;");
-    Assert.assertTrue("expect testConcat node", cg.getNodes(t1m).iterator().hasNext());
+        TypeReference.findOrCreate(ClassLoaderReference.Application, "Lnestmates/Outer$Inner");
+    MethodReference t1m = MethodReference.findOrCreate(t1s, "triple", "()I");
+    Assert.assertTrue("expect Outer.Inner.triple node", cg.getNodes(t1m).iterator().hasNext());
     CGNode t1node = cg.getNodes(t1m).iterator().next();
 
-    // Check call from main to testConcat()
+    // Check call from main to Triple()
     Assert.assertTrue(
-        "should have call site from main to StringConcat.testConcat()",
+        "should have call site from main to TestNestmates.triple()",
         cg.getPossibleSites(mnode, t1node).hasNext());
+
+    // check that triple() does not call an accessor method
+    Assert.assertTrue(
+        "there should not be a call from triple() to an accessor method",
+        cg.getSuccNodes(t1node).hasNext() == false);
   }
 }
